add: layoutsを軸にしたファイル構成に変更
