--- src/app/page.module.css
@@ -1,7 +1,0 @@
-.main {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  padding-top: 6rem;
-}
