+++ src/layouts/Today/index.tsx
@@ -1,0 +1,5 @@
+const Today: React.FC = () => {
+  return <div>ここは/です</div>
+}
+
+export default Today
